Increase C++ compiler warnings level; address the warnings raised by gcc

+++ src/util.h
@@ -30,7 +30,7 @@
 #include <vector>
 #include "constants.h"
 
-#define COUNT_OF(arr) (sizeof(arr) / sizeof(arr[0]))
+#define COUNT_OF(arr) int(sizeof(arr) / sizeof(arr[0]))
 
 inline double signOf(double x) { return x > 0 ? +1 : -1; }
 inline double sqr(double a) { return a * a; }
